prototype/CI/object-detection-service/requirements.txt
fastapi
transformers
uvicorn
opencv-python-headless
pillow
requests